+++ release_notes.md
@@ -1,4 +1,4 @@
-# Azure Functions CLI 4.2.2
+# Azure Functions CLI 4.2.3
 
 #### Host Version
 
@@ -6,5 +6,4 @@
 - In-Proc Host Version: 4.41.100 (4.841.100, 4.641.100)
 
 #### Changes
-
-- Fix .NET template install bug (#4612)
+- Add `func pack` functionality to other languages (#4600)